+++ bothan-binance/src/api.rs
@@ -2,5 +2,6 @@ pub use error::{ConnectionError, Error, SubscriptionError};
 pub use websocket::{BinanceWebSocketConnection, BinanceWebSocketConnector};
 
 pub mod error;
+pub mod msgs;
 pub mod types;
 pub mod websocket;
